+++ src/actions/actionHelper.js
@@ -394,3 +394,19 @@ export function createZoomToFeatureAction(
     },
   };
 }
+
+/**
+ * Creates an action for renaming an item in a VcsList.
+ * @param {VcsListItem} item
+ * @param {ActionOptions} [actionOptions={}]
+ * @returns {VcsAction}
+ */
+export function createRenameAction(item, actionOptions = {}) {
+  return {
+    name: 'list.renameAction.title',
+    callback: () => {
+      item.rename = true;
+    },
+    ...actionOptions,
+  };
+}